Moved Typifier to common
